Исправление замечаний

+++ src/Main.java
@@ -5,11 +5,11 @@ public class Main {
     public static void main(String[] args) {
         Scanner numb = new Scanner(System.in);
         System.out.print("Вес (кг.): ");
-        int weight = numb.nextInt();
+        double weight = numb.nextInt();
         System.out.print("Рост (см.): ");
-        int high = numb.nextInt();
+        double high = numb.nextInt();
         BmiService service = new BmiService();
-        double bmi=service.calculate(high, weight);
+        double bmi = service.calculate(high, weight);
         System.out.printf("Индекс Массы Тела - %f", bmi);
         numb.close();
     }
